% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barlyk.R
\name{barlyk}
\alias{barlyk}
\title{barlyk
#' Trace un graphique en barre des quelques items en Likert scale}
\usage{
barlyk(dfx, titre = "")
}
\arguments{
\item{dfx}{tibble (format long)}

\item{titre}{titre du graphique}
}
\value{
graphique
}
\description{
barlyk
#' Trace un graphique en barre des quelques items en Likert scale
}
\examples{
prepax(likx,1:6) |>  barlyk(titre = "Seances")

}
