% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facetlik.R
\name{facetlik}
\alias{facetlik}
\title{facetlyk
Trace un graphique en barre par ligne de quelques items en Likert scale}
\usage{
facetlik(dfx, titre = "")
}
\arguments{
\item{dfx}{tibble (format long)}

\item{titre}{titre du graphique}
}
\value{
graphique
}
\description{
facetlyk
Trace un graphique en barre par ligne de quelques items en Likert scale
}
\examples{
prepax(likx,1:6) |>  facetlik(titre = "Seances")

}
