% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepax.R
\name{prepax}
\alias{prepax}
\title{PREPAX}
\usage{
prepax(dfx, col)
}
\arguments{
\item{dfx}{data-frame to explore}

\item{col}{numbers of columns to treat}
}
\value{
a tibble
}
\description{
Creat a long tibble with just the selected columns
}
