% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablik.R
\name{tablik}
\alias{tablik}
\title{tablik}
\usage{
tablik(dfx, expx = FALSE, classeur = "", nomtab = "")
}
\arguments{
\item{dfx}{tibble (long format)}

\item{expx}{If true, export on xls}

\item{classeur}{name of the xls document}

\item{nomtab}{name of the sheet}
}
\value{
a table  and a table in xls
}
\description{
Table for some items on likert scale
}
\examples{
prepax(likx,1:6) |>  tablik(expx = FALSE)

}
